+++ docs/source/conf.py
@@ -15,9 +15,10 @@ from recommonmark.parser import CommonMarkParser
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
-# import os
-# import sys
-# sys.path.insert(0, os.path.abspath('.'))
+import os
+import sys
+
+sys.path.insert(0, os.path.abspath("../.."))
 
 
 # -- Project information -----------------------------------------------------
@@ -25,6 +26,7 @@ from recommonmark.parser import CommonMarkParser
 project = "sam"
 copyright = "2018-2022, Royal HaskoningDHV"
 author = "Royal HaskoningDHV"
+
 
 config = toml.load("../../pyproject.toml")